requirements.txt
amqp==2.6.1
wheel
arrow==0.14.7
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
bs4==0.0.1
celery==4.4.7
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.5
django-rest-auth==0.9.5
djangorestframework==3.12.2
flower==0.9.7
gevent==21.1.2
greenlet==1.0.0
humanize==3.2.0
ics==0.7
idna==2.10
kombu==4.6.11
prometheus-client==0.8.0
prompt-toolkit==3.0.14
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.5
redis==3.5.3
requests==2.25.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
TatSu==5.5.0
tornado==6.1
urllib3==1.26.3
vine==1.3.0
wcwidth==0.2.5
zope.event==4.5.0
zope.interface==5.2.0

